[sync] Some work on a large code refactoring towards mk III SDK revision.
Right now only fantasy target, and only some parts of code ported from mk II, mostly for synchronization purposes.

+++ sdk/rckid/graphics/font.h
@@ -31,7 +31,7 @@ namespace rckid {
         int size; 
         int numGlyphs;
         GlyphInfo const * glyphs;
-        uint32_t const * pixels;
+        uint8_t const * pixels;
 
         template<typename T>
         static constexpr Font const fromROM() {
